# redis: move command lookup from map to trie