--- a/source/common/http/header_map_impl.h
+++ b/source/common/http/header_map_impl.h
@@ -110,19 +110,14 @@ protected:
     const LowerCaseString* key_;
   };
 
+  typedef StaticLookupResponse (*EntryCb)(HeaderMapImpl&);
+
   /**
    * This is the static lookup table that is used to determine whether a header is one of the O(1)
    * headers. This uses a trie for lookup time at most equal to the size of the incoming string.
    */
-  struct StaticLookupTable {
+  struct StaticLookupTable : public TrieLookupTable<EntryCb> {
     StaticLookupTable();
-    typedef StaticLookupResponse (*EntryCb)(HeaderMapImpl&);
-
-    void add(const char* key, EntryCb cb) { lookup_table_.add(key, cb); }
-
-    EntryCb find(const char* key) const { return lookup_table_.find(key); }
-
-    TrieLookupTable<EntryCb> lookup_table_;
   };
 
   struct AllInlineHeaders {