--- a/test/extensions/filters/network/redis_proxy/BUILD
+++ b/test/extensions/filters/network/redis_proxy/BUILD
@@ -7,8 +7,8 @@ load(
 )
 load(
     "//test/extensions:extensions_build_system.bzl",
-    "envoy_extension_cc_test",
     "envoy_cc_test_binary",
+    "envoy_extension_cc_test",
 )
 
 envoy_package()