--- a/test/extensions/filters/network/redis_proxy/BUILD
+++ b/test/extensions/filters/network/redis_proxy/BUILD
@@ -7,8 +7,8 @@ load(
 )
 load(
     "//test/extensions:extensions_build_system.bzl",
-    "envoy_cc_test_binary",
     "envoy_extension_cc_test",
+    "envoy_extension_cc_test_binary",
 )
 
 envoy_package()
@@ -94,9 +94,10 @@ envoy_extension_cc_test(
     ],
 )
 
-envoy_cc_test_binary(
+envoy_extension_cc_test_binary(
     name = "command_lookup_speed_test",
     srcs = ["command_lookup_speed_test.cc"],
+    extension_name = "envoy.filters.network.redis_proxy",
     external_deps = [
         "benchmark",
     ],