For an explanation of how to fill out the fields, please see the relevant section 
in [PULL_REQUESTS.md](./PULL_REQUESTS.md)

*Description*: This PR moves the command lookup from unordered_map to trie to generally improve the lookup performance.
*Risk Level*: Low
*Testing*: Existing automated tests and added a speed test
*Docs Changes*: N/A
*Release Notes*: N/A