@ramaraochavali 2019-01-07
@mattklein123 PTAL, when you get time. Here are the benchmark results before and after the change.

**BEFORE (with map)**

```
Benchmark                Time           CPU Iterations
-------------------------------------------------------
BM_MakeRequests    2689920 ns    1563737 ns        434
```
**AFTER (with trie)**

```
-------------------------------------------------------
Benchmark                Time           CPU Iterations
-------------------------------------------------------
BM_MakeRequests    1723733 ns     930528 ns        724
```